Move selfupdate-related services together in a package
