+++ lyrebird/src/test/java/moe/lyrebird/model/update/selfupdate/BinaryInstallationServiceTest.java
@@ -16,7 +16,7 @@
  *     along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package moe.lyrebird.model.update.system;
+package moe.lyrebird.model.update.selfupdate;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
